feat: jwt확인 파일 모듈화

+++ controller/LikeController.js
@@ -1,25 +1,33 @@
 const db = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
-const jwt = require("jsonwebtoken")
-const dotenv = require('dotenv')
-dotenv.config();
+const ensureAuthorization = require('../controller/auth');
+const jwt = require('jsonwebtoken')
 
 const addLike = (req,res)=>{
     const book_id = req.params.id;
     // const {user_id} = req.body;
 
     let authorization = ensureAuthorization(req);
+    if ( authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해주세요"
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "없는 토큰입니다."
+        })
+    } else {
+        let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (1, 3)";
+        let values= [authorization.id, book_id];
+        db.query(sql, values, (err, results)=>{
+            if (err){
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (1, 3)";
-    let values= [authorization.id, book_id];
-    db.query(sql, values, (err, results)=>{
-        if (err){
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-
-        return res.status(StatusCodes.OK).json(results)
-    })                                              
+            return res.status(StatusCodes.OK).json(results)
+        })                                              
+    }
 }
 
 const removeLike = (req,res)=>{
@@ -27,24 +35,26 @@ const removeLike = (req,res)=>{
     // const {user_id} = req.body;
 
     let authorization = ensureAuthorization(req);
+    if ( authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해주세요"
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "없는 토큰입니다."
+        })
+    } else {
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;";
+        let values= [authorization.id, book_id];
+        db.query(sql, values, (err, results)=>{
+            if (err){
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;";
-    let values= [authorization.id, book_id];
-    db.query(sql, values, (err, results)=>{
-        if (err){
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-
-        return res.status(StatusCodes.OK).json(results)
-    })         
-}
-
-function ensureAuthorization(req) {
-    let recivedJWT = req.headers["authorization"];
-    let decodedJWT = jwt.verify(recivedJWT, process.env.SECRTE_KEY)
-
-    return decodedJWT
+            return res.status(StatusCodes.OK).json(results)
+        })
+    }
 }
 
 module.exports = {
